+++ src/product/product.controller.ts
@@ -6,6 +6,8 @@ import {
   Delete,
   Param,
   Body,
+  NotFoundException,
+  BadRequestException,
 } from '@nestjs/common';
 import { ProductService } from './product.service';
 import { Product } from '@prisma/client';
@@ -34,18 +36,14 @@ export class ProductController {
   async getProductById(
     @Param('id') id: string,
   ): Promise<Response<Product | null>> {
-    try {
-      const product = await this.productService.getProductById(id);
-      return {
-        message: 'Successfully retrieved product',
-        data: product,
-      };
-    } catch {
-      return {
-        message: 'Product not found',
-        data: null,
-      };
+    if (!(await this.productService.checkProductExists(id))) {
+      throw new NotFoundException(`Product with ID ${id} not found`);
     }
+    const product = await this.productService.getProductById(id);
+    return {
+      message: `Successfully retrieved product ${product.name}`,
+      data: product,
+    };
   }
 
   // POST: Create a new product using Prisma types
@@ -54,16 +52,20 @@ export class ProductController {
     @Body() productData: Omit<Product, 'id'>,
   ): Promise<Response<Product>> {
     const { categoryId, unitId, storageId, ...rest } = productData;
-    const product = await this.productService.createProduct(
-      rest,
-      unitId,
-      categoryId,
-      storageId,
-    );
-    return {
-      message: 'Product successfully added',
-      data: product,
-    };
+    try {
+      const product = await this.productService.createProduct(
+        rest,
+        unitId,
+        categoryId,
+        storageId,
+      );
+      return {
+        message: `Product ${product.name} successfully added`,
+        data: product,
+      };
+    } catch {
+      throw new BadRequestException('Error creating product');
+    }
   }
 
   // PATCH: Update a product by ID
@@ -72,26 +74,25 @@ export class ProductController {
     @Param('id') id: string,
     @Body() productData: Omit<Product, 'id'>,
   ): Promise<Response<Product>> {
-    try {
-      const product = await this.productService.updateProduct(id, productData);
-      return {
-        message: 'Product successfully updated',
-        data: product,
-      };
-    } catch {
-      return {
-        message: 'Product not found',
-        data: undefined,
-      };
+    if (!(await this.productService.checkProductExists(id))) {
+      throw new NotFoundException(`Product with ID ${id} not found`);
     }
+    const product = await this.productService.updateProduct(id, productData);
+    return {
+      message: `Product ${product.name} successfully updated`,
+      data: product,
+    };
   }
 
   // DELETE: Delete a product by ID
   @Delete(':id')
   async deleteProduct(@Param('id') id: string): Promise<Response<Product>> {
+    if (!(await this.productService.checkProductExists(id))) {
+      throw new NotFoundException(`Product with ID ${id} not found`);
+    }
     const product = await this.productService.deleteProduct(id);
     return {
-      message: 'Product successfully deleted',
+      message: `Product ${product.name} successfully deleted`,
       data: product,
     };
   }
