+++ src/storage/storage.controller.ts
@@ -6,6 +6,8 @@ import {
   Delete,
   Param,
   Body,
+  NotFoundException,
+  BadRequestException,
 } from '@nestjs/common';
 import { StorageService } from './storage.service';
 import { Product, Storage } from '@prisma/client';
@@ -38,39 +40,30 @@ export class StorageController {
   async getStorageById(
     @Param('id') id: string,
   ): Promise<Response<Storage | null>> {
-    try {
-      const storage = await this.storageService.getStorageById(id);
-      return {
-        message: 'Successfully retrieved storage',
-        data: storage,
-      };
-    } catch {
-      return {
-        message: 'Storage not found',
-        data: null,
-      };
+    if (!(await this.storageService.checkStorageExists(id))) {
+      throw new NotFoundException(`Storage with ID ${id} not found`);
     }
+    const storage = await this.storageService.getStorageById(id);
+    return {
+      message: `Successfully retrieved storage ${storage.name}`,
+      data: storage,
+    };
   }
 
   @Get(':id/products/')
   async getProductsInStorage(
     @Param('id') id: string,
   ): Promise<Response<Product[]>> {
-    try {
-      const storage = await this.storageService.getStorageById(id, {
-        products: true,
-      });
-      return {
-        message:
-          'Successfully retrieved al products in storage ' + storage.name + '.',
-        data: storage.products,
-      };
-    } catch {
-      return {
-        message: 'Storage not found',
-        data: null,
-      };
+    if (!(await this.storageService.checkStorageExists(id))) {
+      throw new NotFoundException(`Storage with ID ${id} not found`);
     }
+    const storage = await this.storageService.getStorageById(id, {
+      products: true,
+    });
+    return {
+      message: `Successfully retrieved all products in storage ${storage.name}.`,
+      data: storage.products,
+    };
   }
 
   // POST: Create a new storage using Prisma types
@@ -78,11 +71,15 @@ export class StorageController {
   async createStorage(
     @Body() storageData: Omit<Storage, 'id'>,
   ): Promise<Response<Storage>> {
-    const storage = await this.storageService.createStorage(storageData);
-    return {
-      message: 'Storage successfully added',
-      data: storage,
-    };
+    try {
+      const storage = await this.storageService.createStorage(storageData);
+      return {
+        message: `Storage ${storage.name} successfully added`,
+        data: storage,
+      };
+    } catch {
+      throw new BadRequestException('Error creating storage');
+    }
   }
 
   // PATCH: Update a storage by ID
@@ -91,26 +88,25 @@ export class StorageController {
     @Param('id') id: string,
     @Body() storageData: Omit<Storage, 'id'>,
   ): Promise<Response<Storage>> {
-    try {
-      const storage = await this.storageService.updateStorage(id, storageData);
-      return {
-        message: 'Storage successfully updated',
-        data: storage,
-      };
-    } catch {
-      return {
-        message: 'Storage not found',
-        data: undefined,
-      };
+    if (!(await this.storageService.checkStorageExists(id))) {
+      throw new NotFoundException(`Storage with ID ${id} not found`);
     }
+    const storage = await this.storageService.updateStorage(id, storageData);
+    return {
+      message: `Storage ${storage.name} successfully updated`,
+      data: storage,
+    };
   }
 
   // DELETE: Delete a storage by ID
   @Delete(':id')
   async deleteStorage(@Param('id') id: string): Promise<Response<Storage>> {
+    if (!(await this.storageService.checkStorageExists(id))) {
+      throw new NotFoundException(`Storage with ID ${id} not found`);
+    }
     const storage = await this.storageService.deleteStorage(id);
     return {
-      message: 'Storage successfully deleted',
+      message: `Storage ${storage.name} successfully deleted`,
       data: storage,
     };
   }
